+++ src/exchanges/cn/huobi/XHuobi.cs
@@ -372,7 +372,7 @@ namespace CCXT.Simple.Exchanges.Huobi
 
                                 _ticker.volume24h = Math.Floor(_volume / mainXchg.Volume24hBase);
 
-                                var _curr_timestamp = DateTimeExtensions.NowMilli;
+                                var _curr_timestamp = TimeExtensions.NowMilli;
                                 if (_curr_timestamp > _next_timestamp)
                                 {
                                     _ticker.volume1m = Math.Floor((_prev_volume24h > 0 ? _volume - _prev_volume24h : 0) / mainXchg.Volume1mBase);
@@ -457,7 +457,7 @@ namespace CCXT.Simple.Exchanges.Huobi
 
                             _ticker.volume24h = Math.Floor(_volume / mainXchg.Volume24hBase);
 
-                            var _curr_timestamp = DateTimeExtensions.NowMilli;
+                            var _curr_timestamp = TimeExtensions.NowMilli;
                             if (_curr_timestamp > _next_timestamp)
                             {
                                 _ticker.volume1m = Math.Floor((_prev_volume24h > 0 ? _volume - _prev_volume24h : 0) / mainXchg.Volume1mBase);