+++ src/exchanges/cn/kucoin/XKucoin.cs
@@ -132,7 +132,7 @@ namespace CCXT.Simple.Exchanges.Kucoin
 
         public void CreateSignature(HttpClient client, string endpoint)
         {
-            var _timestamp = DateTimeExtensions.NowMilli.ToString();
+            var _timestamp = TimeExtensions.NowMilli.ToString();
 
             var _sign_data = _timestamp + "GET" + endpoint;
             var _sign_hash = Convert.ToBase64String(Encryptor.ComputeHash(Encoding.UTF8.GetBytes(_sign_data)));
@@ -420,7 +420,7 @@ namespace CCXT.Simple.Exchanges.Kucoin
 
                             _ticker.volume24h = Math.Floor(_volume / mainXchg.Volume24hBase);
 
-                            var _curr_timestamp = DateTimeExtensions.NowMilli;
+                            var _curr_timestamp = TimeExtensions.NowMilli;
                             if (_curr_timestamp > _next_timestamp)
                             {
                                 _ticker.volume1m = Math.Floor((_prev_volume24h > 0 ? _volume - _prev_volume24h : 0) / mainXchg.Volume1mBase);
@@ -507,7 +507,7 @@ namespace CCXT.Simple.Exchanges.Kucoin
 
                             _ticker.volume24h = Math.Floor(_volume / mainXchg.Volume24hBase);
 
-                            var _curr_timestamp = DateTimeExtensions.NowMilli;
+                            var _curr_timestamp = TimeExtensions.NowMilli;
                             if (_curr_timestamp > _next_timestamp)
                             {
                                 _ticker.volume1m = Math.Floor((_prev_volume24h > 0 ? _volume - _prev_volume24h : 0) / mainXchg.Volume1mBase);